Fix clang-cl build, benchmark flags, and topological sort stack overflow
- cmake: guard /MP flag to cl.exe only (clang-cl doesn't support it)
- cmake: add clang-cl-specific warning suppressions (-Wno-unknown-attributes,
  -Wno-unused-parameter/variable, propagate CLANG_WARNINGS -Wno-* flags)
- benchmark: add required 's' suffix to --benchmark_min_time (v1.9.1 API)
- topological_sort: convert recursive DFS to iterative to fix stack overflow
  on deep graphs (e.g. 10000-vertex path DAG in BM_TopoSort benchmark)
- Broad warning cleanup across views, algorithms, generators, io, tests,
  and examples for clang-cl compatibility (unused params, unqualified std::move/forward, etc.)

+++ tests/container/dynamic_graph/test_dynamic_graph_cpo_unordered.cpp
@@ -1149,7 +1149,7 @@ TEMPLATE_TEST_CASE("unordered CPO target(g, uv)", "[dynamic_graph][cpo][target]"
     std::vector<uint32_t> targets;
     for (auto uv : edge_view) {
       auto target_vertex = target(g, uv);
-      targets.push_back(vertex_id(g, target_vertex));
+      targets.push_back(static_cast<uint32_t>(vertex_id(g, target_vertex)));
     }
     std::ranges::sort(targets);
     REQUIRE(targets.size() == 2);
@@ -1219,7 +1219,7 @@ TEMPLATE_TEST_CASE("unordered CPO target(g, uv)", "[dynamic_graph][cpo][target]"
     // Unordered: collect targets
     std::vector<uint32_t> targets;
     for (auto uv : edge_view) {
-      targets.push_back(vertex_id(g, target(g, uv)));
+      targets.push_back(static_cast<uint32_t>(vertex_id(g, target(g, uv))));
     }
     std::ranges::sort(targets);
     REQUIRE(targets.size() == 2);
