+++ tests/views/test_transpose.cpp
@@ -81,9 +81,9 @@ TEST_CASE("transpose_view - vertex CPOs forward unchanged", "[views][transpose]"
   // Verify vertices range produces same vertex ids
   std::vector<uint32_t> g_ids, tv_ids;
   for (auto&& [uid, u] : view::vertexlist(g))
-    g_ids.push_back(uid);
+    g_ids.push_back(static_cast<uint32_t>(uid));
   for (auto&& [uid, u] : view::vertexlist(tv))
-    tv_ids.push_back(uid);
+    tv_ids.push_back(static_cast<uint32_t>(uid));
 
   REQUIRE(g_ids == tv_ids);
 }
@@ -236,3 +236,4 @@ TEST_CASE("transpose_view - single vertex no edges", "[views][transpose]") {
 
   CHECK(adj_list::num_vertices(tv) == 1);
 }
+
